webapp_backend/requirements.txt
Flask==1.1.2
flask-restful==0.3.8
requests==2.24.0
psycopg2-binary==2.8.6
flask-sqlalchemy==2.5.1
Flask-Restless==0.17.0
flask-cors